make pvn-wrapper a command with sub commands
