+++ go.mod
@@ -1,3 +1,10 @@
 module github.com/prodvana/pvn-wrapper
 
 go 1.20
+
+require github.com/spf13/cobra v1.7.0
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
